Make the range match stop on , and }
The range match is fairly greedy, and has a high priority.
nft doesn't restrict it to numbers on both sides so I'm not going to,
but it does not accept things that could be interpreted as a string such
as ssh-radius (which is an acceptable range if written as "ssh -
radius") while this syntax does.

This reduces greediness of the range match as a workaround for #3, and
allows the `string - string` syntax so that correct input is properly
classified.

TODO: correctly parse `string-string` as a string instead of a range